Limit the validating of destination address to the actual number of nodes
Currently, validating the dump destination address, on each node, is done
using a hard coded "max nodes per system" enumeration, which is currently
set at 8.

This becomes problematic for a single node system, where checks are done on non
existing nodes which eventually causes an error.

The Perl script processMrw.pl accurately calculates the max nodes per system and
stores this data in attribute ATTR_MAX_COMPUTE_NODES_PER_SYSTEM. Using this
value in lieu of the hard coded enum ensures that only the nodes that are
actually on the system get validated and therefore avoids the error when
validating on non-existing nodes.

Change-Id: Ie99d7340c0729de062f8778502b878449c7e2190
CQ: SW554389
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134468
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>