PRD: single node CS isolation when non-SMP coherent
During the IPL, before all of the nodes are stitched together, each node
will IPL independently. If one of these nodes reports a system checkstop
in this mode, error isolation must remain scoped to that node. It was
previously thought that the list of chips at attention given to PRD from
ATTN would only contain chips in the checkstopped node. However, that is
not the case. Therefore, we needed to restore some checks that we pulled
from the P10 fabric sorting code back in 2020 so that isolation remains
scoped to the node with the checkstop.

Change-Id: Ib8a84955203afdaaaf005773cbc796da91c8f558
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134648
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134921
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>