Hostboot OFFLINE PEL parsing data_files
Fix the duplicate data_files which was caused by out-of-order changes
to the layout of the setup parameters.

Commit 0b06a4afbfb771ba797775f028c6975f1b81e18e
Jun 3 2022 caused a duplicate data_files argument to be added.

Commit 3c369d94b3ef3eaba58e50fea9d7de9d7a75d5e2
Jul 12 2022 modified the data_files.

The above two commits modified same data structure, but the order
of the parameters in the argument list allowed both to be added,
where the intent was to modify the single data_files argument.

Change-Id: I97ae3b25c3e57e89ea4a4cf100519c774afc98fb
CQ: SW554899
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/135208
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>