Inhibit reconfig loops due to deconfigs through host guard
If an early SBE error occurs, FSI init step may attempt to process
the resulting error log and deconfigure the processor, causing a reconfig
loop which will defeat Hostboot's ability to process that early failure
and invoke an SBE dump.  To avoid that, this commit prevents reconfig
loops due to deconfigurations in the isteps prior to discover targets.
That unifies the behavior for all isteps through end of host guard.
As part of this rework, created a new attribute to control whether
a deconfig is able to trigger a reconfig loop, and created a helper function
to set/clear different reconfig loop reasons.

Change-Id: I53bf412ece8c16a460cb3e8c2056239ea0f7f234
CQ: SW554244
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/135473
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>