Log error when rcToErrl fails to create error log
This commit adds FFDC changes to log an error for cases where
rcToErrl is not able to translate a FAPI return code into a
valid platform error log, indicating a firmware bug.  This
will help determine which error path is hitting in the
transition driver SBE update path and, until now, causing
a crash.

Change-Id: I266f3ad7fd2425e877a3e34812d4184887f67292
CQ: SW554464
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/137263
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>