Clear "Deconfigured By EID" field when re-enabling FCO cores
On a MPIPL, cores that were previously deconfigured due to FCO might
be called on to fill in for another core that was deconfigured.  In
this case Hostboot needs to reset/clear some ATTR_HWAS_STATE fields.
Previously, the "Deconfigured By EID" field was not being cleared, but
this change now clears that field.

Change-Id: Ife9ee40d2152609d2e79846832455e3c797b1f05
CQ:SW553882
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/137458
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>