Save every committed log to a file in simics
In standalone simics there is limited space in memory for all of
the PELs we create, especially when running all of the testcases.
This change introduces a new magic instruction that will save the
complete PEL contents to a file when it is committed.  The file
name will be "pel.<EID>.bin".

Change-Id: I4df2bb1c6241d8aa14e7275e3535a16a7019a72e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/137727
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>