Fixes Explorer clear row repair bad bits byte position bug
In Explorer, the MCBIST/CCS engines report the ECC and spare bytes
as bytes 8-9. The DFI reports the ECC and spare bytes as bytes 4-5.
The clear row repair function did not properly take into account the
different bus perspectives.

This commit addds in a helper function to swizzle from the MCBIST/CCS
byte perspective to the DFI byte perspective. It adds this swizzle to
the clear row repairs function

ZEN:MST-1721
CQ:SW555654
Change-Id: I4811f881f8c1727a067d7a9647c274cc147c1b05
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/137824
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN AMMANDA DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/137831
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>