eBMC setup script: Fix config file verification
Fixes a bug in the CONFIG_FILE verification where if the user already
had an existing p10ebmc.config file in their repo (from a previous build
setup or otherwise) the script would assume all is correct. However, to
ensure there is no possibility that the existing config file is from
another build an md5sum is run against the existing config file and the
one from the build. If there is a mismatch then the existing is
overridden.

Change-Id: I6e74dfbceef70af1e51ea1ff6e1664af4f6d3e18
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138104
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138364