Fix "FW Released Ver" Field in Hostboot PELs
This commit fixes a mistake in a previous commit that overlooked
FRU VPD that hostboot receives from the BMC with "FW_VERSION_INFO".
With this fix the PELs are now properly using information from the
Marker Lid's Mi Keyword for the "FW Released Ver" field.

For example, on a recent fw1030 driver that value was set to
"NL1030_008", and is the same value that the BMC and PHYP components
use for their PELs.

Change-Id: I6eed54c89c8e6563e0bcc81e086b7c9022c3f02f
CQ:SW556590
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138256
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>