Reboot instead of TI when mismatch PDRs are found
At the end of the PDR exchange, Hostboot compares the PDRs given by the
BMC with what it was given initially. If the amount given at the end of
the exchange was less than what was given initially, then Hostboot would
emit an error resulting in a TI.

With this change, Hostboot will emit an informational log and perform a
reboot. This allows for recovery in the case that during the early IPL a
customer pulls out their LCD Display

Change-Id: I7d0934dd0c43fac467871ff59e86e53f67cbb165
EWM:PE00DBKX
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138571
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138620
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>