Add Common Fail Messages to HBBL Console Output
This commit adds extra traces to the Hostboot Bootloader (HBBL)
console output for the 2 most common secureboot failures: fails
on different HW Keys' Hashes and fails when the incoming container
(usually the Hostboot Base Code - HBB) has a secure version
less than the minimum secure version of the system.

The extra traces in the console should make it easier for people
to diagnose the problem and resolve it faster.

Change-Id: Idb24fe789fd3a2d197f139122d9311de28d472b6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138708
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>