Don't allow clock test failure logs to fail the istep
Previously, when a clock test failed in the presence of a redundant
clock, Hostboot would downgrade the severity of the error log, add a
reference to the hwp log to the istep error log (which would cause the
istep to fail), and request a reconfig loop. However, the process of
adding the hwp log to the istep log would upgrade the severity of the
hwp log again, which violates the design specification.

This commit makes a change so that in that scenario, the code requests
a reconfig loop and commits the hwp error log, but does not add the
hwp error log to the istep error log. This will make the istep
"succeed" and perform a reconfig loop without upgrading the severity
of the hwp error log.

Change-Id: Id1d805ddeee90ea8a755490346efec2330fbc0a2
CQ: SW556431
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/138737
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>