PNOR Alternate Master Part Callouts
The ErrlEntry::setGardType and ErrlEntry::setDeconfigState methods
will now allow new parameters which will identify the type and style
of the callout(s) desired to be updated.

This allows explicit action based on the caller's input, as well as
retaining the legacy functionality of performing updates to ALL of
the supported callout types, i.e. HW_CALLOUT, CLOCK_CALLOUT and
PART_CALLOUT.

Additionally, platHandlePartCallout needs to use the appropriate structure
fields to properly mark the data.

Change-Id: I83a8ab2c1de947f4e435e74670969c51475b1b37
CQ:  SW556591
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/139581
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>