Fix fail in checkMinHw when node0 is guarded
There was an issue on multi-node systems where if the primary node was
guarded for any reason then the minimum hardware check would fail
despite other potential boot procs being available on other nodes.

To fix this issue, when getting a handle to the boot proc during the
minimum hardware check we now request that it be functional instead of
purely following the blueprint. Without requesting a functional proc
back from the function the algorithm to find the boot proc will always
return the lowest topology id boot proc which could be non-functional.

Change-Id: I9d757a5f7bacb33b7a4caa3aa1929b4fa99503c1
EWM:SW555084
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/139946
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>