PRD: check for two bad clocks in integrated spare mode
Everest system are configured with integrated spare clocks. We only want
to make a serviceable log when both clocks are bad. Our previous design
assumed that if both clocks are bad then the system will checkstop. This
is only the case after the SMP buses have been trained. Before then, the
PLLs can generate psuedo clocks from its VCO even if the input clock is
dead. So, there is a window early in the IPL where both clocks can be
dead but the system will not checkstop. Therefore, we will need to add
code into Hostboot PRD to check if both clocks are bad in this
integrated spare configuration.

Change-Id: Ia1d945fe3e56ae7c12e72cae9e776f2525206c03
CQ: SW557225
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140048
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140116
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140178
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>