Allow inline CXX execution
Modified TestSuite.H so that it can be included in arbitrary
code and allow execution of individual testcases.

Include the header for your particular testcase at the top
of your .C (e.g. one of the istep functions)
 #define MANUAL_CXXTEST
 #include "../../testcore/pir/test_pirformat.H"

Then instantiate a test class within your function and call
the test function(s) that you want to execute.
 PirformatTest test1;
 test1.testPirformat();

Output will show up in traces as usual.

Change-Id: If7ca5860a20d80da323423b903e7f337f6651b37
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/143342
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>