Changes for Persistent Attr Updates
Problem Statement:
We want to detect any persistent attributes (non-volatile, RW) being
changed after istep16.

Solution Implemented:
Introduced a new flag that will be checked in _trySetAttr().
This flag will be set to false at the end of istep 16 to
indicate we do not want to allow persistent attrs to be updated.
The _trySetAttr will check the flag and if it indicates we
cannot update persistent attr, it will get the metadata
corresponding to the attr being updated, and will add an
error log entry if its persistent. However it will allow the
attr to be updated even if the new flag is set and the
attr being changed is persistent. The error log entry will
help us if this issue occurs. If the _trySetAttr is not
able to find the metadata associated with the attr, it
will add an error log entry to detect this case.

For the testcases to succeed, the flag will be set to
allow updates to persistent attribute just before we start
the CXX_testcases.

Change-Id: Ic690e4e74099478c0e96482a1b7a2febb03604a9
JIRA:PFHB-299
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/139955
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>