Get SPD from the correct EEPROM_SOURCE for isdimms
When calling cmpEecacheToEeprom() on isdimms, we call readFromEepromSource()
to compare the Cache and HW values, which eventually call spdFetchData() to
perform the data read. However, this call to spdFetchData() doesn't specify
the EEPROM_SOURCE and defaults to AUTOSELECT. This will then only do the data
reads from the EECACHE and will incorrectly pass the comparison when control
returns to cmpEecacheToEeprom(). This is bad as any changes to either EECACHE
or EEPROM HW will be seen as 'in sync' when they could be completely different,
allowing HB to ipl with different EECACHE and EEPROM HW data. This commit
fixes this by passing in the EEPROM_SOURCE to spdFetchData() from the caller.

Additionally, if a difference between the EECACHE and EEPROM HW on isdimms,
we should not perform CRC checking. Instead, we will assume new HW and
refresh the cache with the HW data.

Change-Id: Ia3b3e8c74bd377db09b4f0d3f580ee7ff7103290
JIRA:PFHB-12
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140980
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>