update_subtree.sh: add required option for newer git versions
Starting with git v2.10 the 'git merge' command used in this tool to
merge all the latest changes for a subtree requires the option
'--allow-unrelated-histories'. Otherwise, the command will fail with
error message "fatal: refusing to merge unrelated histories".

The shebang was changed to run as a bash script because it has bash
specific syntax that is not posix compliant.

A check was added to verify if the user is alread in an 'hb workon'
before sourcing 'env.bash'.

Change-Id: I01c88c8251eb5a1f5a4bc98a712cab3683de63b9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140483
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140494
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>