update_subtree.sh: handling merge conflicts with newer git
There is some code to handle merge conflicts after merging the lastest
from upstream. This currently uses grep and awk to pull some information
out of git status. However, on newer versions of git the status format
changed such that the awk was pulling from the wrong position. I changed
this to use the --porcelain option which is portable among git versions.

In addition, changed the script to only try to resolve the conflicts if
they exist. Previously, the script would throw an error and move on.

Change-Id: I02eabe89d407ab36023490c07226c3c121a2b08e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140536
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140733
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>