update_subtree.sh: trouble determining last synched commit
The current code uses a gerrit query over ssh to get the commit ID of
the last synched commit. The query searches for the Change-Id and does
not check for the correct branch which is problematic if the commit has
been cherry-picked to multiple branches. In addition, the query will
fail on a newly created subtree because 'git subtree add' creates a
merge commit that has to bypass gerrit reviews.

Since we already look in the current working tree to get the Change-Id
of the last synched commit, the last synched commit should also be in
the current working directory. So instead of dealing with the
problematic gerrit query, the tool was changed to get the commit ID of
the last synched commit from the current woring directory.

Change-Id: I4a0d316b9b9259a29da200b4e329fbdd1e88b193
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140539
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140732
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>