update_subtree.sh: exit if no new changes from upstream
The current script continues to process things even if there are no new
changes from the upstream subtree repo. Fortunately, it fails for the
most part with out any impacts, but the error messages are confusing. So
this tool was modified to exit cleanly if there are no new changes.

Also, supressed error message when trying to remove non-existent patch
files.

Change-Id: I1a41429259a000125a109cfa4d24f7ce3cb304d0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140550
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140734
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>