Runtime support for page switch ops on ONE_BYTE_ADDR_PAGESELECT EEPROMS
For non-PAGESELECT EEPROMs reads and writes can be performed to any
given offset within the device at anytime. However, PAGESELECT EEPROMs
need to be aware of what page they are currently working with, and switch
pages if necessary. This switch is done by writing 2 bytes to one of two
special device addresses (0x6C or 0x6E). This commit ports the page switch
behavior from ipl code to runtime to detect if a page switch was requested
and set up the special device address for the following runtime code to
write to.

Change-Id: If44b5e3b608833c3f5ea23001d759be380bd5ac2
JIRA:PFHB-23
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140546
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>