EEPROM redundancy fixes for Planar OCMBs and isdimms
EEPROM_VPD_BACKUP_INFO does not exist for the isdimms and
Planar OCMBs modeled in the Bonito xml, however EEPROM writes
are still attempted to their (non-existant) BACKUP copy. To
correct this, this commit updates the xml to include
EEPROM_VPD_REDUNDANCY_NOT_PRESENT for isdimms, POSSIBLE for
lcard-dimm-ddimms, and NOT_POSSIBLE for chip-ocmb-planar.
Also changes spdUpdateEepromRedundancy() to check for
EEPROM_VPD_BACKUP_INFO on dimms if they are not DDIMMs
to set redundancy PRESENT/NOT_PRESENT appropriately.

Change-Id: I99dd2f3b1fde08be1f74bdc41256964f288b4d68
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140687
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>