update_subtree.sh: add upstream revision input
The current tool will only update to the latest master. However, that
may not always be desired. An option has now been provided to allow the
user to input a specific upstream commit ID. If not provided the default
is to used the master branch.

A side benefit that this is new option can also be used to revert back
to a previous revision if the current revision is too far forward.

Change-Id: Ia32919eee7fd810adbf282a484de74fb476bbba7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140571
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/140738
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>