setup-script: Add choice to clobber repo post setup
There is now a choice to clobber the hostboot repo and standalone
sandbox to complete the setup procedure. Initially this was not part of
the setup script to allow users more flexibility in how they proceed
post script execution. Now, a yes/no prompt will allow the user to
decide if the script should handle it for them.

Change-Id: I07a2135ebc114dda5963069031e40d0ac28ceb9d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/141163
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>