Add Capability for PNOR::PSPD Overlay
Provide the ability to build an SPD image into the
PNOR::PSPD partition.

For OCMB's soldered onto the backplane, allow the ability
to build an SPD image to be released in future firmware
releases.

Hostboot will detect if the PNOR::PSPD partition should be
used (if available) and ignore the originally shipped
backplane VPD SPD.

Change-Id: I41c0eaeb59565eb00317deae4f3dc79a6df5f446
JIRA: PFHB-5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/141182
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>