Add Exceptions for OCMB SPD Overlay Logic to Retain Some Backplane Data
Perform range checks that will restore any special bytes needing to be
retained from the original Backplane VPD.  The only data that needs to
be retained currently is DMB Revision (aka OCMB DD levels) and the
RAW_CARD_BYTES_OFFSET (DMB Revision is to properly identify the
chip level that is currently installed).

Currently the only identified bytes are used in the byte array.

In the future the byte array can be expanded in order to allow
more restoration of the Backplane VPD (if desired or required
in the future).

More Context - The PNOR content is a post-GA overlay of the
memory characterization data that exists inside the backplane
VPD EEPROM.  The purpose of the overlay is to either tweak the
settings based on learning or add support for new ISDIMM types.

Change-Id: Ib904d0185a9444d60d22bbf1e840212adb69257d
JIRA: PFHB-5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/141462
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>