Only interrogate boot processor for SBE SMP integration capability
Hostboot queries the boot processor SBE to determine whether it
can perform the nodal SMP integration, saving the result
into an attribute.  However, there is a bug such that when Hostboot
moves on to query the secondary processors for their respective SBE
capabilities, which may differ from the boot processor, it erroneously
overwrites the attribute again, potentially causing the SMP integration
procedure to perform the wrong SMP integration flow.  This change
prevents any updates to the attribute unless the SBE in question is the
boot processor.

Change-Id: Ibbc2102c08f6fe7c78fbd4ea275e6b6be21c3084
STGD: 388819
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/141726
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>