PRD: Remove "PHYP running" check in TOD analysis
The P9 code assumes there is no reason for PRD FSP to ever analyze a TOD
error. Unfortunately, there is a use case were TOD errors can cause a TI
attention. Adjustments need to be made so that FSP PRD can safely
analyze the TOD error. This is a backport of P10 changes made in SW533268,
which enabled TOD error analysis for checkstop attentions.

Change-Id: Ic40c2419132bb95ce1c0a14e94c7462328652a82
STGD: 402203
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/144765
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/145076
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>