Support for MCTP tagging
Modify Hostboot to send, receive, and process MCTP tags
as part of MCTP communications with the BMC.  Hostboot
will send a new 7-bit MCTP tag for each logical message
it sends to the BMC, incrementing the tag value for
each new message, and looping back to 0 once it
increments past the maximum tag value.  The MCTP layer
will ensure it repeats the same tag value for each packet
of a logical message.  If a PLDM message times out and
Hostboot retries the transation, it will assign a new tag
since the retry is logically a new MCTP message.  The
same behavior is expected when the remote end transmits
MCTP messages to hostboot.  In addition to setting the
tag value, Hostboot also sets the "tag owner" field to
1 whenever it initiates a message, and sets it to 0
when responding to a BMC request.

Change-Id: I4fe614b014dee6b0f26fa34091355b30db4a9a27
JIRA: PFHB-308
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/145064
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>